requirements.txt
pygame==2.0.0.dev6
pyyaml
Pillow
pymunk
pyglet
numpy
